+++ LineFile.h
@@ -1,6 +1,6 @@
 #pragma once
 struct LineFile {
-	int data_size = 11;
+	short data_size = 11;
 	char data[11];
-	double open, high, low, close;
+	float open, high, low, close;
 };
